+++ src/main/java/com/axperty/stackedblocksfarmersdelight/StackedBlocksFarmersDelight.java
@@ -7,12 +7,13 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 public class StackedBlocksFarmersDelight implements ModInitializer {
-    public static final String MODID = "stackedblocksfarmersdelight";
-    public static final Logger LOGGER = LoggerFactory.getLogger(MODID);
+    public static final String MOD_ID = "stackedblocksfarmersdelight";
+    public static final Logger LOGGER = LoggerFactory.getLogger(MOD_ID);
 
     @Override
     public void onInitialize() {
         CreativeTabRegistry.registerItemGroups();
         BlockRegistry.registerModBlocks();
+        LOGGER.info("Stacked Blocks: Farmer's Delight loaded");
     }
 }